Changed from struct to class
Made a change from struct to class.
copy constructors still need work, ao and streams too. Will need to do the
destructors and check for memory leaks soon.

I started on the driver. I should have everything needed to do the driver. Not sure how I'll handle the play method later on. I can either pinpoint which card the player is trying to play or just play any card of the type the player wants that is in his hand. It shouldn't matter for the demo as they only mention that all the cards in the hand should be player