play() is broken :(
The method play() is broken.
I currently believe that it is broken because the iterator cannot find the pointer in the hand of cards. This is most likely due to it because 2 different pointers.

I need to come up with a fix for it. Either by looking for the Card object directly or by making sure that it is the same Card* everywhere.

*** I added some getters to Player and I tested the driver, everything still works fine. ***